Module Two Assignment
For this assignment, you will create the pattern of a chess board that
is 8 x 8.  Use X and O to represent the squares.

Create the appropriate nested looping structure to output the characters
in an 8 x 8 grid on the screen using Console.Write() or
Console.WriteLine() as appropriate.
Include a decision structure to ensure that alternate rows start with
opposite characters as a real chess board alternates the colors among
rows.
This is what your output should look like.

XOXOXOXO
OXOXOXOX
XOXOXOXO
OXOXOXOX
XOXOXOXO
OXOXOXOX
XOXOXOXO
OXOXOXOX